+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Improved documentation
+- Added runnable example illustrating crate usage
 
 
 0.1.0